faster-rcnn & mask-rcnn train and test support

+++ mmdet/models/bbox_heads/bbox_head.py
@@ -109,7 +109,7 @@ class BBoxHead(nn.Module):
             # TODO: add clip here
 
         if rescale:
-            bboxes /= scale_factor.float()
+            bboxes /= scale_factor
 
         if nms_cfg is None:
             return bboxes, scores
